Fix  #2569 - fix variant export
Avoid 2 variants with same name
This will not fix older files, we need to manually update your variant names
