+++ addons/io_scene_gltf2/blender/com/gltf2_blender_utils.py
@@ -78,3 +78,24 @@ def fast_structured_np_unique(arr, *args, **kwargs):
         return (unique,) + result[1:]
     else:
         return unique
+
+def find_unused_name(haystack, desired_name):
+    """Finds a name not in haystack and <= 63 UTF-8 bytes.
+    (the limit on the size of a Blender name.)
+    If a is taken, tries a.001, then a.002, etc.
+    """
+    stem = desired_name[:63]
+    suffix = ''
+    cntr = 1
+    while True:
+        name = stem + suffix
+
+        if len(name.encode('utf-8')) > 63:
+            stem = stem[:-1]
+            continue
+
+        if name not in haystack:
+            return name
+
+        suffix = '.%03d' % cntr
+        cntr += 1
